refactor(nuget): split nuget.rs into nested submodule per ADR 0009
Split the 3,031-line nuget.rs into a nuget/ directory with surface files
following the two-scale module structure prescribed by ADR 0009. Each
independently-registered PackageParser and its private helpers live in a
dedicated surface file, with shared helpers and types in mod.rs.

Surface files:
- packages_config.rs, nuspec.rs, packages_lock.rs, deps_json.rs,
  project_json.rs, project_file.rs, directory_props.rs, nupkg.rs, utils.rs

Test files moved into nuget/ as co-located submodules:
- nuget_test.rs, nuget_scan_test.rs

The golden test (nuget_golden_test.rs) also moved into nuget/ but remains
loaded via #[path] from parsers/golden_test.rs to avoid duplicate module
loading with the existing golden test infrastructure.
