+++ src/parsers/nuget/nuget_golden_test.rs
